feat(update-bar): add script to update the configuration of the bar process
The `update-bar` script is added to the project. This script is used to update the configuration of the bar process. It retrieves the path of the `bar.sh` script, sets the `current_config` variable to "onedark 2", and sends the `SIGUSR1` signal to the bar process using `pkill`. This allows for dynamic configuration updates without restarting the bar process.