chore(nvim_lazyvim): add dap plugins and align.nvim plugin to lazyvim.json
The lazyvim.json file was modified to add the "lazyvim.plugins.extras.dap.core" and "lazyvim.plugins.extras.dap.nlua" plugins to the extras array. This change was made to enable support for debugging with the DAP protocol in Neovim.

The lazyvim.json file was also modified to add the "lazyvim.plugins.extras.editor.aerial" plugin to the extras array. This change was made to enable the Aerial plugin, which provides a code outline view in the editor.

The coding.lua file was modified to update the filetypes table. The previous individual entries for markdown, yaml, gitcommit, and gitrebase were replaced with a wildcard entry ["*"] = true. This change was made to enable the coding plugin to work with any filetype.

The editor.lua file was modified to add the "tpope/vim-abolish" plugin to the plugins array. This change was made to add support for the vim-abolish plugin, which provides enhanced text manipulation capabilities.

The editor.lua file was also modified to add the "Vonr/align.nvim" plugin to the plugins array. This change was made to add support for the align.nvim plugin, which provides alignment functionality for code and text.

feat(zsh): add aliases for bun and bunx commands

The aliases.zsh file was modified to add the "bn" and "bx" aliases for the "bun" and "bunx" commands, respectively. These aliases were added to provide convenient shortcuts for running the "bun" and "bunx" commands.

feat(zsh): add ya function

The aliases.zsh file was modified to add the "ya" function. This function is used to run the "yazi" command with a specified working directory. The function creates a temporary file to store the current working directory, runs the "yazi" command with the specified arguments and the "--cwd-file" option pointing to the temporary file, and then changes the current working directory to the one specified in the temporary file. This function was added to simplify the process of running the "yazi" command with a specific working directory.

chore(zsh): source ~/.fzf.zsh if it exists

The bin_settings.zsh file was modified to check if the ~/.fzf.zsh file exists and source it if it

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>